refactor(claw-server): split HTTP routes and handlers
