+++ packages/browseros-agent/apps/claw-server/src/api/http/errors.ts
@@ -1,4 +1,6 @@
 import type { ApiError } from '@browseros/claw-api'
+import type { Context } from 'hono'
+import type { RequestContextEnv } from './request-context'
 
 /**
  * Error envelope for the canonical API surface. Route handlers and the
@@ -13,4 +15,13 @@ export function canonicalApiError(
   return requestId === undefined
     ? { code, message }
     : { code, message, requestId }
+}
+
+export function apiError(
+  c: Context<RequestContextEnv, string>,
+  status: 400 | 403 | 404 | 409 | 410 | 413,
+  code: string,
+  message: string,
+) {
+  return c.json(canonicalApiError(code, message, c.get('requestId')), status)
 }
